Add tutorial model, combined main model
The tutorial model contains Sequence, a list of Step objects. For now,
Step is just a String, but that will expand. Each Step is a piece of the
tutorial for the user.

The tutorial element is hidden if there are no items in the sequence.
The plan is to reload the initial sequence when it's done, but if it
turns out that it would be good to turn it off in some cases, this
provides a way to do that.