Message, update for tutorial
Currently there are two messages in the tutorial, and just one
message for advancing. The update function is structured to allow
Tutorial.update to see the whole model, which will be useful for
controlling the state to guide interactions.

Tutorial.elm owns the wrapped model, but Main.elm owns the wrapping
view function. This split is different than what I've done elsewhere,
like in SpaceCommand, but here it helps with keeping the header
defined in Main.elm.