Update tutorial max width
I experimented with some long messages, and found that this max width
is good for taking up most of the top of the screen. Also, the text can
wrap to two lines without affecting the vertical spacing. It should be
possible to keep the messages within this width while still allowing
for a good amount of content.