Fix tutorial cache resetting every step
Prior to this, the cache was loaded from the Space state every step,
meaning that with multiple steps before restoring from the cache,
the user's original state was lost. The state from the previous tutorial
step was used instead.

Now, the cache is only loaded from the Space state if it is empty,
meaning that the user's original state is preserved. The restore step
resets the cache to its initial empty state so that it will be filled
with the user's state if triggered again.