Specify widths with flex
The command bar now stays at a constant 300px width, with the space
taking up the rest of the width.

Currently the height stretches past the bottom of the screen if the
window is broad enough, because its aspect ratio stays the same, making
the user scroll to see the whole image. This will need to be the next
improvement.