Handle 'choice' dropdown with onChange
The new default message is needed for the function to compile, because
the onChange event receives a string value and then has to select the
correct message from the list. There's no non-maybe way to do that
selection. At least by defining the DoNothing message, any bizarre
behavior should be avoided in the near-impossible event this is actually
triggered.

I verified the workflow in Firefox and Chromium. This fixes #13.