Rewrite dropdown fix using Json.Decode.fail
Thanks to helpful conversation on the Elm Discourse, I realized that
`Json.Decode.fail` throws a runtime error in the way that `Events.on`
expects. Handling the matching of the value to the message within the
`onChangeDiscrete` function allows for handling the default case without
passing in a dummy message.