Wrap baseContents in UndoList
The only state that undo/redo will control is baseContents, because
all the other pieces have to do with display aspects and are easily
toggled. The base shape and the IterFrames are the things that have
fine controls a user might want to manage with the undo stack.

UndoList provides a built in update wrapper, but it looks like it calls
new with every wrapped message. This would not be ideal for pointer
events; using undo to go to one of thousands of previous drag points
would be painful. Instead, undo should take the user to the end of the
previous drag. Getting this to work requires implementing the handling.

Two calls to `UndoList.new` are added here because they seemed obvious:
- removing an IterFrame
- hiding shapes during the tutorial
    (may not be used, but no reason not to have this option)