Make iteration frame controls visible at all times
Toggling the visibility of the iteration frames on the main space used
to also toggle the visibility of the iteration frame controls in the
command bar. This didn't add much value, and it made the layout
inconsistent.

Part of why I want to fix this is so that I can move the canvas state
to the bottom of the command bar. The button to show the encoding
will live there.