Parse bookmark from URL parameter
This was a big effort with lots of trial and error. Some decisions to
highlight:

Because URL events must lead to messages, but they could be duds, Main
needs to have a message value that means no change. Main is the only
module that needs an exception for doing nothing, so it owns
representing this by wrapping the message type in a maybe.

I tried matching to URLs like '/bookmark', but regular http servers
return a 404 because they don't find a file at a path there. It might be
possible to configure a server to redirect that to the app, but I don't
have control over the server on github.io, and that seems like not a
great requirement generally. We'll just have to live with the query
syntax.

Surprisingly, you can match on a query parameter with an empty string
like this: fp/?=EQ . Hey, it will save a character.

Matchers for paths that include index.html are included. elm reactor
produces paths like this, so this should be supported.

The URL change events are not currently in use. Entering a url in the
browser always triggers a page load from init. Still, I may want to
allow alternate ways of triggering them eventually, so it seems best to
leave these wired up.