Encode and decode iteration frames
The iteraction frame specific tests check point similarity after
applying to a fuzzed point, which is the best way to verify that they
will look the same on the screen. I did relax the point similarity in
the tests to 0.2 pixels to account for extra error accumulation in the
transform. This is still plenty accurate for the screen.

Implementing the decoder requires passing the reference frame through
awkwardly.  I'm finally convinced that I should change the way that the
reference frame is tied in to the IterFrame view cycle. That can wait
for later, though. The main consequence is that the reference frame on
initial page load is always used when the bookmark is loaded from the
URL, regardless of what it was when it was made.

The encoding for the Sierpinski carpet takes up 154 base 64 characters.
While that's a lot, it's serviceable for most social media platforms,
so I'm taking it as a win!